requirements/dev.txt
-r base.txt

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-cov>=4.1
factory-boy>=3.3

# Debug
django-debug-toolbar>=4.3
ipython>=8.20

# Code quality
ruff>=0.3
